refacto: déplacement des classes BDESE dans un fichier spécifique
