+++ impact/reglementations/forms/bdese.py
@@ -4,7 +4,7 @@ from django import forms
 from django.core.exceptions import ValidationError
 from django.db import models
 
-from reglementations.models import CategoryType
+from reglementations.models.bdese import CategoryType
 from utils.forms import DsfrForm
