+++ impact/reglementations/tests/test_bdese_forms.py
@@ -7,8 +7,8 @@ from reglementations.forms.bdese import BDESE_300_FIELDS
 from reglementations.forms.bdese import BDESE_50_300_FIELDS
 from reglementations.forms.bdese import bdese_configuration_form_factory
 from reglementations.forms.bdese import bdese_form_factory
-from reglementations.models import BDESE_300
-from reglementations.models import BDESE_50_300
+from reglementations.models.bdese import BDESE_300
+from reglementations.models.bdese import BDESE_50_300
 
 
 @pytest.fixture
